deleted unwanted files and moved some components
