--- src/App1.scss
@@ -1,267 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: "Lato", sans-serif;
-  width: 100%;
-  a {
-    text-decoration: none;
-    color: white;
-  }
-  a :hover {
-    text-decoration: none;
-  }
-}
-
-.sidenav {
-  height: 100%;
-  width: 160px;
-  position: fixed;
-  z-index: 1;
-  top: 0;
-  left: 0;
-  background-color: #fff;
-  overflow-x: hidden;
-  padding-top: 20px;
-  font-family: "Slabo 27px", serif;
-  border-right: 1px solid gray;
-}
-
-.sidenav li {
-  padding: 6px 8px 6px 16px;
-
-  font-size: 1.5rem;
-  color: #818181;
-  display: block;
-}
-
-.sidenav li:hover {
-  color: #5f5858;
-  cursor: hand;
-  text-decoration: none !important;
-}
-
-.main {
-  margin-left: 160px; /* Same as the width of the sidenav */
-  font-size: 28px; /* Increased text to enable scrolling */
-  height: 51px;
-  display: flex;
-}
-
-.Logo img {
-  width: 100px;
-  height: 100px;
-  margin-left: 20px;
-}
-.Logo {
-  text-align: center;
-}
-
-.aside {
-  text-align: center;
-  margin-left: 170px;
-  margin-top: 10px;
-}
-
-h3 {
-  text-align: center;
-}
-.home {
-  margin-top: 70px;
-  margin-left: -15%;
-  text-align: center;
-  height: 300px;
-}
-.homeImage {
-  margin: 10px auto 530px;
-  width: 200px;
-  height: 200px;
-}
-@media screen and (max-height: 450px) {
-  .sidenav {
-    padding-top: 15px;
-  }
-  .sidenav li {
-    font-size: 18px;
-  }
-}
-
-form.title input[type="text"] {
-  padding: 12px;
-  font-size: 17px;
-  border: 1px solid transparent;
-  background: #f1f1f1;
-  width: 80%;
-}
-
-form.title::after {
-  content: "";
-  clear: both;
-  display: table;
-}
-form.title {
-  margin-left: 160px;
-  margin-bottom: 40px;
-  margin-top: 40px;
-}
-.editor {
-  width: 100%;
-  position: relative;
-}
-
-.listgroupItem > a {
-  color: black;
-  text-decoration: none !important;
-}
-.contentList {
-  margin-top: 20px;
-  width: fit-content;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: stretch;
-  max-height: fit-content;
-}
-.btn-group {
-  margin: auto;
-}
-.topnav {
-  background-color: #3366d2;
-  overflow: hidden;
-  width: 100%;
-  display: flex;
-}
-.topnav span {
-  text-decoration: none !important;
-  color: #f2f2f2;
-  padding: 10px 10px;
-  font-size: 17px;
-  position: fixed;
-  right: 0px;
-}
-.topnav span > a {
-  padding: 10px 10px;
-}
-.topnav span > a:hover {
-  background-color: #ddd;
-  color: black;
-  text-decoration: none;
-}
-#myTopnav {
-  position: absolute;
-  display: none;
-  height: auto;
-  padding: 0 10px;
-  right: 3px;
-  top: 3px;
-  z-index: 1;
-}
-@media screen and (max-width: 600px) {
-  #myTopnav {
-    display: block;
-  }
-  .homeAndAdd {
-    display: flex;
-    flex-direction: column;
-  }
-}
-
-.search {
-  position: relative;
-  color: #aaa;
-  font-size: 16px;
-}
-.search input {
-  margin-top: 5px;
-  margin-left: 10px;
-  width: 250px;
-  height: 38px;
-  background: #fcfcfc;
-  border: 1px solid #aaa;
-  border-radius: 5px;
-  box-shadow: 0 0 3px #ccc, 0 10px 15px #ebebeb inset;
-}
-
-.search input {
-  text-indent: 10px;
-}
-.search .fa-search {
-  position: absolute;
-  top: 2px;
-  right: 1px;
-  font-size: 25px;
-  z-index: 1;
-  color: rgb(105, 101, 101);
-}
-
-.listTitles {
-  width: 250px;
-  height: 100px;
-  font-size: 16px;
-  margin-left: 10px;
-  background-color: black;
-  color: white;
-  position: fixed;
-  z-index: 1;
-  overflow: scroll;
-  display: none;
-}
-.titles {
-  display: block;
-}
-.topnav {
-  overflow: hidden;
-  background-color: #333;
-}
-
-.topnav a {
-  float: left;
-  display: block;
-  color: #f2f2f2;
-  text-align: center;
-  padding: 14px 16px;
-  text-decoration: none;
-  font-size: 17px;
-}
-
-.topnav a:hover {
-  background-color: #ddd;
-  color: black;
-}
-
-.topnav a.active {
-  background-color: #4caf50;
-  color: white;
-}
-
-.topnav .icon {
-  display: none;
-}
-
-@media screen and (max-width: 600px) {
-  .topnav a:not(:first-child) {
-    display: none;
-  }
-  .topnav a.icon {
-    float: right;
-    display: block;
-  }
-}
-
-@media screen and (max-width: 600px) {
-  .topnav.responsive {
-    position: relative;
-  }
-  .topnav.responsive .icon {
-    position: absolute;
-    right: 0;
-    top: 0;
-  }
-  .topnav.responsive a {
-    float: none;
-    display: block;
-    text-align: left;
-  }
-}
